!B fixed warning

+++ Program.cpp
@@ -65,7 +65,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdLine
 		}
 
 		auto currentTime = std::chrono::high_resolution_clock::now();
-		float deltaTimeInSeconds = ((double)std::chrono::duration_cast<std::chrono::microseconds>(currentTime - lastTime).count()) * 0.001 * 0.001;
+		float deltaTimeInSeconds = (float)(((double)std::chrono::duration_cast<std::chrono::microseconds>(currentTime - lastTime).count()) * 0.001 * 0.001);
 		lastTime = currentTime;
 
 		gEngine->Update(deltaTimeInSeconds);
